add nav bar
