+++ src/__tests__/Home/MainFeedScreen.test.js
@@ -1,7 +1,6 @@
 import React from 'react'
 import { shallow, mount } from 'enzyme'
 import { useSelector, useDispatch } from 'react-redux'
-import axios from 'axios'
 import MainFeedScreen from '../../main/Home/MainFeedScreen'
 import { getImagesAction } from '../../main/Image/imageApi'
 import { getPromptsAction } from '../../main/Prompt/promptsApi'
@@ -26,8 +25,6 @@ const mockState = {
   user,
 }
 const dispatch = jest.fn()
-
-const GATEWAY_URL = process.env.REACT_APP_GATEWAY_URL
 
 describe('<MainFeedScreen>', () => {
   let wrapper
@@ -57,80 +54,6 @@ describe('<MainFeedScreen>', () => {
   })
 
   describe('<div> for appContainer', () => {
-    describe('user', () => {
-      it('should render user email', () => {
-        expect(wrapper.find({ 'data-testid': 'user' }).text()).toEqual(
-          'SomeUser',
-        )
-      })
-    })
-    describe('Logout Button', () => {
-      it('should render with correct text', () => {
-        expect(wrapper.find({ 'data-testid': 'logoutButton' }).text()).toEqual(
-          'Logout',
-        )
-      })
-      it('dispatch with correct param when clicked', () => {
-        wrapper.find({ 'data-testid': 'logoutButton' }).simulate('click')
-        expect(dispatch).toHaveBeenCalledWith({ type: TYPES.LOGOUT })
-      })
-    })
-    describe('Add Prompts Button', () => {
-      describe('when user has role GODLIKE', () => {
-        const godlikeUser = {
-          id: 'some id',
-          email: 'SomeUser',
-          role: 'GODLIKE',
-        }
-        beforeEach(() => {
-          jest.resetAllMocks()
-          useSelector.mockImplementation(callback =>
-            callback({ user: godlikeUser }),
-          )
-        })
-        it('should render with correct text', () => {
-          const newWrapper = shallow(<MainFeedScreen />)
-          expect(newWrapper.find({ testid: 'promptButton' }).text()).toEqual(
-            'Add Prompts',
-          )
-        })
-        it('calls prompts api when clicked', () => {
-          const newWrapper = shallow(<MainFeedScreen />)
-          newWrapper.find({ testid: 'promptButton' }).simulate('click')
-          expect(axios.post).toHaveBeenCalledWith(
-            `${GATEWAY_URL}/api/prompts?userId=${godlikeUser.id}`,
-          )
-        })
-      })
-      describe('when user has role FEEDER', () => {
-        it('should not render add prompt button', () => {
-          jest.resetAllMocks()
-          const feederUser = {
-            id: 'some id',
-            email: 'SomeUser',
-            role: 'FEEDER',
-          }
-          useSelector.mockImplementation(callback =>
-            callback({ user: feederUser }),
-          )
-          const newWrapper = shallow(<MainFeedScreen />)
-          expect(newWrapper.find({ testid: 'promptButton' })).toHaveLength(0)
-        })
-      })
-    })
-    describe('header', () => {
-      it('renders with the correct text', () => {
-        expect(wrapper.find({ 'data-testid': 'header' }).text()).toEqual(
-          'Daily Art Prompt',
-        )
-      })
-      it('is red', () => {
-        expect(
-          wrapper.find({ 'data-testid': 'header' }).prop('style'),
-        ).toEqual({ color: 'red' })
-      })
-    })
-
     describe('<PromptLayout>', () => {
       it('should exist', () => {
         expect(wrapper.find({ 'data-testid': 'promptLayout' })).toHaveLength(1)
