+++ src/main/Home/MainFeedScreen.js
@@ -1,5 +1,4 @@
 import React, { useEffect } from 'react'
-import axios from 'axios'
 import { useDispatch, useSelector } from 'react-redux'
 import './main.css'
 import PromptLayout from '../Prompt/PromptLayout'
@@ -8,11 +7,11 @@ import { getPromptsAction } from '../Prompt/promptsApi'
 import * as TYPES from '../storage/actions'
 import ImageGallery from '../Image/ImageGallery'
 import ImageUpload from '../Image/ImageUpload'
+import DapNavBar from './DapNavBar'
 
 const MainFeedScreen = () => {
   const user = useSelector(state => state.user)
   const dispatch = useDispatch()
-  const GATEWAY_URL = process.env.REACT_APP_GATEWAY_URL
 
   useEffect(() => {
     dispatch(getImagesAction(user.id))
@@ -22,30 +21,7 @@ const MainFeedScreen = () => {
 
   return (
     <div data-testid="appContainer" className="app">
-      <div data-testid="user">{user?.email}</div>
-      <button
-        data-testid="logoutButton"
-        onClick={() => dispatch({ type: TYPES.LOGOUT })}
-      >
-        Logout
-      </button>
-      {user.role === 'GODLIKE' && (
-        <button
-          testid="promptButton"
-          onClick={() =>
-            axios.post(`${GATEWAY_URL}/api/prompts?userId=${user.id}`)
-          }
-        >
-          Add Prompts
-        </button>
-      )}
-      <div className="header">
-        <div className="title">
-          <h1 data-testid="header" style={{ color: 'red' }}>
-            Daily Art Prompt
-          </h1>
-        </div>
-      </div>
+      <DapNavBar />
       <PromptLayout data-testid="promptLayout" />
       <hr />
       <h1 data-testid="artGalleryHeader" className="title">
