+++ src/main/Home/main.css
@@ -64,3 +64,11 @@ img {
   vertical-align: super;
   color: red;
 }
+
+#nav-user-dropdown {
+  color: white;
+}
+
+#nav-user-dropdown::after {
+  display: none;
+}
